+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - PIM-9780: Fix completed import/export job notification broken link
 - PIM-9715: Prevent the deletion of an attribute used as a label by a family
 - PIM-9781: Fix Category tree not refreshing when switching locale
+- PIM-9739: Fix connection users, users, channels having a link to a sub-category
 
 ## New features